envsetup: mka,reposync *Add the PureNexus specific functions (mka, reposync)
Change-Id: Ieb5a3c80c2a447b6986fc028389532b9a92f59d0
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>

envsetup.sh: Fix the `reposync` function

It seems ionice can't find `repo` if it's in ~/bin, even if ~/bin
is in the user's $PATH.  Placing repo into /usr/bin or /usr/local/
bin works, but is undesirable (build instructions from both AOSP
and CM tell users to put repo into ~/bin, and repo likes to update
itself from time to time).

This forces the reposync function to use the full path to the repo
binary.

Change-Id: I9dc4a3d4ec3f39563e3a84de7321189700079c2e
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>