build: condition A53 errata workaround in RenderScript
* following the logic I introduced in c546185 and 18e9c5e .
* CPU targets which are not similar to Cortex-A53 don't need this.
-> since using the workarounds results in a performance pentalty,
   disable them for targets which don't need them.

Change-Id: I8dff81617f51dd9d63614f9c0306d793dfe6e259
Signed-off-by: Alex Naidis <alex.naidis@linux.com>

arm64: Rework A53 Errata Fixing logic

This commit introduces an internal variable
"APPLY_A53_ERRATA_FIXES" which will be the final
indicator on whether to apply the Errata fixes
for Cortex-A53 or not. Since the variable is internal
only, it gets cleared before and after usage.

Additionally to the CPU detection logic, we are
now reading "TARGET_IS_CORTEX-A53" which is a flag
that CAF sets to true on chipsets which contain
a Cortex-A53, even if they are using "generic"
as their CPU target. So this is the only way
to find out whether an Cortex-A53 is really
present.

We are overriding any results from the CPU
target detection with what comes from
"TARGET_IS_CORTEX-A53" if it is present now.

That means that, we are also introducing a
way to force opting in and force opting out of
the Cortex-A53 Errata fixes.

Additionally the CPU detection logic
is deduplicated and cleaned up.

Change-Id: I1f46cdc41087c4b45c365211be5e9ef0bbdc97f1
Signed-off-by: Alex Naidis <alex.naidis@linux.com>